+++ changelog.md
@@ -302,7 +302,7 @@
     * Added check_machine_name_availability
     * Added check_farm_name_availability
     * Added check_rds_server_name_availability
-    * 
+
 07-04-2021
 * Config Class
     * Added get_im_assets
@@ -311,3 +311,7 @@
     * Added get_im_tags
     * Added get_im_versions
     * Added get_im_version
+
+10-04-2021
+* Config Class
+    * Added new_im_asset